Added so that twarc tests work with stream
