+++ sc/tw/.gitignore
@@ -6,3 +6,4 @@
 project
 target
 tmp
+twarc.log
