--- sc/tw/src/test/scala/resources/.gitignore
@@ -1,2 +1,0 @@
-*.jsonl
-!testTweets.jsonl
